@@ -1,7 +1,7 @@
 [package]
 name = "nix-ld"
 version = "2.0.4"
-edition = "2021"
+edition = "2024"
 
 [[bin]]
 name = "nix-ld"
@@ -11,13 +11,17 @@ test = false
 name = "tests"
 
 [dependencies]
-cfg-match = "0.2.1"
-constcat = "0.6.1"
-default-env = "0.1.1"
 embedded-io = "0.6.1"
-goblin = { version = "0.9.3", default-features = false, features = ["elf32", "elf64"] }
+goblin = { version = "0.9.3", default-features = false, features = [
+    "elf32",
+    "elf64",
+] }
 heapless = "0.8.0"
-linux-raw-sys = { version = "0.9.4", default-features = false, features = ["no_std", "general", "errno"] }
+linux-raw-sys = { version = "0.9.4", default-features = false, features = [
+    "no_std",
+    "general",
+    "errno",
+] }
 log = "0.4.27"
 
 [build-dependencies]
@@ -25,7 +29,6 @@ cc = "1.2.25"
 
 [dev-dependencies]
 cc = "1.2.25"
-lazy_static = "1.5.0"
 rstest = { version = "0.25.0", default-features = false }
 tempfile = "3.20.0"
 
